meky

+++ database/migrations/2024_03_26_223907_create_lessons_table.php
@@ -16,14 +16,14 @@ return new class extends Migration
             $table->string('name');
 
             $table->unsignedBigInteger('unit_id')->constrained('units')->cascadeOnDelete();
-        
-            $table->integer('price');
+
+            $table->float('price');
             $table->text('description');
             $table->string('image')->nullable();
             $table->string('file')->nullable();
             $table->string('video')->nullable();
 
-            
+
             $table->timestamps();
         });
     }
